copy full knative file
