+++ internal/kgateway/agentgatewaysyncer/conversion.go
@@ -855,10 +855,6 @@ func filteredReferences(parents []routeParentReference) []routeParentReference {
 	return ret
 }
 
-func getDefaultName(name string, kgw *gwv1.GatewaySpec) string {
-	return fmt.Sprintf("%v-%v", name, kgw.GatewayClassName)
-}
-
 // IsManaged checks if a Gateway is managed (ie we create the Deployment and Service) or unmanaged.
 // This is based on the address field of the spec. If address is set with a Hostname type, it should point to an existing
 // Service that handles the gateway traffic. If it is not set, or refers to only a single IP, we will consider it managed and provision the Service.
